feature: keep only needed pages
