+++ .eslintrc
@@ -90,7 +90,7 @@
         "react/jsx-props-no-spreading": 0,
         "react/require-default-props": 0,
         "no-unused-expressions": 0,
-        "no-unused-vars": 0,
+        "no-unused-vars": 2,
         "jsx-a11y/no-static-element-interactions": 0,
         "jsx-a11y/click-events-have-key-events": 0,
         "jsx-a11y/no-noninteractive-element-interactions": 0,
