+++ src/components/contact/index.module.scss
@@ -5,12 +5,15 @@
 .area {
     @include breakpoint(min-xl-device) {
         padding-top: 140px;
+        padding-bottom: 140px;
     }
     @include breakpoint(lg-device) {
         padding-top: 100px;
+        padding-bottom: 100px;
     }
     @include breakpoint(max-md-device) {
         padding-top: 50px;
+        padding-bottom: 50px;
     }
 }
 .img {
