+++ src/components/layout/offcanvas-data.js
@@ -1,101 +1,24 @@
 export const OffcanvasData = [
     {
-        id: 1,
+        id: 0,
         title: 'Home',
         path: '/',
         cName: 'offcanvas-text',
     },
     {
-        id: 2,
+        id: 1,
         title: 'About',
         path: '/about',
         cName: 'offcanvas-text',
     },
     {
-        id: 3,
-        title: 'Services',
-        path: '/services',
-        cName: 'offcanvas-text',
-    },
-    {
-        id: 4,
+        id: 2,
         title: 'Projects',
-        path: '#',
+        path: '/projects',
         cName: 'offcanvas-text',
-        arrowDown: 'FaAngleDown',
-        submenu: [
-            {
-                id: 'submenu-01',
-                link: '/projects/project-fullwidth',
-                text: 'Project Fullwidth',
-            },
-            {
-                id: 'submenu-02',
-                link: '/projects/project-2-columns',
-                text: 'Two Columns',
-            },
-            {
-                id: 'submenu-03',
-                link: '/projects/project-gallery',
-                text: 'Project Gallery',
-            },
-            {
-                id: 'submenu-04',
-                link: '/projects/project-slider',
-                text: 'Project Slider',
-            },
-        ],
     },
     {
-        id: 5,
-        title: 'Pages',
-        path: '#',
-        cName: 'offcanvas-text',
-        arrowDown: 'FaAngleDown',
-        submenu: [
-            {
-                id: 'submenu-01',
-                link: '/our-clients',
-                text: 'Our Clients',
-            },
-            {
-                id: 'submenu-02',
-                link: '/our-team',
-                text: 'Our Team',
-            },
-            {
-                id: 'submenu-03',
-                link: '/our-working-process',
-                text: 'Our Working Process',
-            },
-        ],
-    },
-    {
-        id: 6,
-        title: 'Blogs',
-        path: '#',
-        cName: 'offcanvas-text',
-        arrowDown: 'FaAngleDown',
-        submenu: [
-            {
-                id: 'submenu-01',
-                link: '/blogs/blog-fullwidth',
-                text: 'Blog FullWidth',
-            },
-            {
-                id: 'submenu-02',
-                link: '/blogs/blog-leftsidebar',
-                text: 'Blog Left Sidebar',
-            },
-            {
-                id: 'submenu-03',
-                link: '/blogs/blog-rightsidebar',
-                text: 'Blog Right Sidebar',
-            },
-        ],
-    },
-    {
-        id: 7,
+        id: 3,
         title: 'Contact',
         path: '/contact',
         cName: 'offcanvas-text',
