+++ src/components/projects/gallery-grid.js
@@ -5,12 +5,8 @@ import classes from './project.module.scss';
 import ProjectGalleryItem from './gallery-item';
 
 function ProjectGalleryGrid({ projects }) {
-    const [noOfElement, setNoOfElement] = useState(3);
+    const [noOfElement] = useState(9);
     const slice = projects.slice(0, noOfElement);
-
-    const loadMore = () => {
-        setNoOfElement(noOfElement + noOfElement);
-    };
 
     return (
         <div className={classes.project}>
@@ -23,24 +19,6 @@ function ProjectGalleryGrid({ projects }) {
                         />
                     ))}
                 </Row>
-                {noOfElement < projects.length && (
-                    <div className={classes.project_btn__wrap}>
-                        <button
-                            type="button"
-                            className={classes.loadmore_btn}
-                            onClick={loadMore}
-                        >
-                            Load more
-                        </button>
-                    </div>
-                )}
-                {noOfElement > projects.length && (
-                    <div className={classes.project_btn__wrap}>
-                        <span className={classes.loadedText}>
-                            All item has been loaded!
-                        </span>
-                    </div>
-                )}
             </Container>
         </div>
     );
