--- src/pages/projects/project-fullwidth.js
@@ -1,58 +1,0 @@
-import Head from 'next/head';
-import PropTypes from 'prop-types';
-import BannerFive from '../../components/banner/index-5';
-import Breadcrumb from '../../components/breadcrumb';
-import Footer from '../../components/layout/footer';
-import Newsletter from '../../components/newsletter/newsletter';
-import ProjectFullwidth from '../../components/projects/fullwidth';
-import { getAllItems } from '../../lib/items-util';
-
-function ProjectFullwidthPage({
-    projects,
-    bannerTwoItems,
-    newsletterItems,
-    footerItems,
-}) {
-    return (
-        <>
-            <Head>
-                <title>Projects FullWidth - OxyBuild</title>
-                <meta name="description" content="Projects FullWidth" />
-            </Head>
-            <Breadcrumb
-                subTitle="Our Project"
-                title="FullWidth"
-                desc="Construction of itself, because it is pain some proper style design occur are pleasure"
-            />
-            <ProjectFullwidth projects={projects} />
-            <BannerFive bannerTwoItems={bannerTwoItems} />
-            <Newsletter newsletterItems={newsletterItems} />
-            <Footer footerItems={footerItems} />
-        </>
-    );
-}
-
-export function getStaticProps() {
-    const allItems = getAllItems('projects');
-    const bannerTwoItems = getAllItems('banner-2');
-    const newsletterItems = getAllItems('newsletter');
-    const footerItems = getAllItems('footer');
-
-    return {
-        props: {
-            projects: allItems,
-            bannerTwoItems,
-            newsletterItems,
-            footerItems,
-        },
-    };
-}
-
-ProjectFullwidthPage.propTypes = {
-    projects: PropTypes.instanceOf(Object).isRequired,
-    bannerTwoItems: PropTypes.instanceOf(Object).isRequired,
-    newsletterItems: PropTypes.instanceOf(Object).isRequired,
-    footerItems: PropTypes.instanceOf(Object).isRequired,
-};
-
-export default ProjectFullwidthPage;
